+++ src/main/java/com/coniverse/dangjang/domain/intro/dto/IntroInfo.java
@@ -1,0 +1,4 @@
+package com.coniverse.dangjang.domain.intro.dto;
+
+public record IntroInfo(String minVersion, String latestVersion) {
+}
