+++ src/main/java/com/coniverse/dangjang/global/config/TimeConfig.java
@@ -1,0 +1,22 @@
+package com.coniverse.dangjang.global.config;
+
+import java.util.TimeZone;
+
+import org.springframework.context.annotation.Configuration;
+
+import jakarta.annotation.PostConstruct;
+
+/**
+ * 서버 시간 동기화
+ *
+ * @author Teo
+ * @since 1.0
+ */
+@Configuration
+public class TimeConfig {
+
+	@PostConstruct
+	public void setTimeZone() {
+		TimeZone.setDefault(TimeZone.getTimeZone("Asia/Seoul"));
+	}
+}
