+++ src/main/java/com/coniverse/dangjang/global/exception/BadRequestException.java
@@ -1,0 +1,11 @@
+package com.coniverse.dangjang.global.exception;
+
+/**
+ * @author Teo
+ * @since 1.0
+ */
+public class BadRequestException extends BusinessException {
+	public BadRequestException() {
+		super(400, "잘못된 요청입니다.");
+	}
+}
