--- src/main/java/dangjang/challenge/ChallengeApplication.java
@@ -1,15 +1,0 @@
-package dangjang.challenge;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@EnableJpaAuditing
-@SpringBootApplication
-public class ChallengeApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ChallengeApplication.class, args);
-	}
-
-}
