--- src/main/java/dangjang/challenge/domain/intro/dto/IntroInfo.java
@@ -1,4 +1,0 @@
-package dangjang.challenge.domain.intro.dto;
-
-public record IntroInfo(String minVersion, String latestVersion) {
-}
