--- src/main/java/dangjang/challenge/global/config/TimeConfig.java
@@ -1,22 +1,0 @@
-package dangjang.challenge.global.config;
-
-import java.util.TimeZone;
-
-import org.springframework.context.annotation.Configuration;
-
-import jakarta.annotation.PostConstruct;
-
-/**
- * 서버 시간 동기화
- *
- * @author Teo
- * @since 1.0
- */
-@Configuration
-public class TimeConfig {
-
-	@PostConstruct
-	public void setTimeZone() {
-		TimeZone.setDefault(TimeZone.getTimeZone("Asia/Seoul"));
-	}
-}
