--- src/main/java/dangjang/challenge/global/support/BaseEntity.java
@@ -1,31 +1,0 @@
-package dangjang.challenge.global.support;
-
-import java.time.LocalDateTime;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-
-/**
- * Entity 공통 매핑 정보
- *
- * @author Teo
- * @since 1.0
- */
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public class BaseEntity {
-	@CreatedDate
-	@Column(name = "CREATED_AT", updatable = false)
-	private LocalDateTime createdAt;
-
-	@LastModifiedDate
-	@Column(name = "MODIFIED_AT")
-	private LocalDateTime modifiedAt;
-}
