+++ src/test/java/com/coniverse/dangjang/DangjangApplicationTests.java
@@ -1,0 +1,13 @@
+package com.coniverse.dangjang;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DangjangApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
