--- src/test/java/dangjang/challenge/ChallengeApplicationTests.java
@@ -1,13 +1,0 @@
-package dangjang.challenge;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ChallengeApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
